+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.3.0 / 2014-03-19
+
+* Add prepare callback and update start callback.
+* Add $.fn.vendorCss(), $.saveStyle(), $.restoreStyle() methods.
+* Add empty keyframes.
+* Add tile plugin.
+* Update plugins.
+
 # v0.2.4 / 2014-03-11
 
 * Fixed slide and fly plugin bug.